Customer Service Requests Analysis
Perform a service request data analysis of New York City 311 calls. You will focus on the data wrangling techniques to understand the pattern in the data and also visualize the major complaint types.
Domain: Customer Service

For more details/tasks performed in the project, please read the "README" file in the repository.